moving functions to helpers.py
